fix every dashboard header
im in pain

+++ admin_dashboard_delete.php
@@ -5,13 +5,19 @@
 	$model = new Model;
 
 	$id = $_GET['id'];
-	$delete = $model->deleteBlog($id);
+    if (isset($_SESSION['role']) && $_SESSION['role'] == 1) {
+	    $delete = $model->deleteBlog($id);
+    }else{
+        $delete = false;
+    }
 
     if($delete){
 
         echo "<script>alert('delete successfully');</script>";
         echo "<script>window.location.href = 'admin_dashboard_blog.php';</script>";
 
+    }else{
+        echo "<script>alert('You're not suppose to do this');</script>";
     }
 
 ?>
